python/pydantic-ai/simple/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
distro==1.9.0
eval_type_backport==0.2.2
Flask==3.1.0
griffe==1.7.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
logfire-api==3.14.0
MarkupSafe==3.0.2
openai==1.75.0
opentelemetry-api==1.32.1
pydantic==2.11.3
pydantic-ai-slim==0.1.3
pydantic-graph==0.1.3
pydantic_core==2.33.1
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
